pyproject.toml
[project]
name = "orca-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jeff Fischer", email = "jeff.fischer@benedat.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[openai]>=0.3.2",
    "requests>=2.32.4",
    "slack-sdk>=3.35.0",
]

[project.scripts]
orca-agent = "orca_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=9.3.0",
    "pytest>=8.4.1",
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
